+++ electron/main.ts
@@ -1,4 +1,5 @@
 import { app, BrowserWindow, nativeImage, ipcMain } from 'electron';
+import * as remoteMain from '@electron/remote/main';
 
 import * as path from 'path';
 
@@ -11,6 +12,8 @@ import Store from 'electron-store';
 import { APP_PROTOCOL_NAME } from '../src/config/StaticConfig';
 
 import { getGAnalyticsCode, getUACode, actionEvent, transactionEvent, pageView } from './UsageAnalytics';
+
+remoteMain.initialize();
 
 (global as any).actionEvent = actionEvent;
 (global as any).transactionEvent = transactionEvent;
@@ -101,13 +104,15 @@ function createWindow() {
     webPreferences: {
       webviewTag: true,
       nodeIntegration: true,
+      nodeIntegrationInSubFrames: true,
       devTools: isDev,
-      enableRemoteModule: true,
       contextIsolation: false,
     },
     resizable: true,
     icon: iconImage,
   });
+
+  remoteMain.enable(win.webContents);
 
   // Note that all efforts to hide menus only work on Windows and Linux
   // The option Menu.setApplicationMenu(null) seemed to have worked on all platforms, but it had some breaking behaviors
@@ -169,7 +174,6 @@ app.on('activate', async () => {
 });
 
 app.on('ready', async function () {
-  app.allowRendererProcessReuse = false;
   createWindow();
 
   await new Promise(resolve => setTimeout(resolve, 20_000));