+++ src/pages/dapp/browser/types.ts
@@ -1,5 +1,4 @@
-import { WebviewTag } from 'electron';
 
-export type WebView = WebviewTag & HTMLWebViewElement;
+export type WebView = Electron.WebviewTag & HTMLWebViewElement;
 
 export type ErrorHandler = (reason: string) => void;